pyproject.toml
[tool.poetry]
name = "hospital-management-system"
version = "0.1.0"
description = ""
authors = ["Akshay <78297310+webintellectual@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "10.0.1"
mysql-connector-repackaged = "0.3.1"
mysql-connector-python = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
